Add test case using pseudo elements & width attr
Add a test case that uses the inlinePseudoElements and
applyWidthAttributes together at the same time. This is a copy of the
pseudo-elements test case with just the applyWidthAttributes option
added to the json file.

This is currently failing due to a bug with the reference to the crated
pseudo element being a jQuery object instead of a raw document element.