Make the "style" attribute name configurable
Working on the client-side and using jQuery you get unexpected results
if you work on "known" attributes like "style" because the browser tries
to understand the CSS and does its own transformations. In order to
preserve the styles created by juice the strategy is to work in a
different attribute and then do a search/replace of the attribute name
once you get the output html (no more in the browser DOM).