Fixed counter-reset not working in combination with inline styles
An error is thrown in `resetCounter` when adding an `counter-reset` inline style to an element in combination with a different style (e.g., in head).

This error is caused by the fact that the `addProps` method is called when an element doesn't have `styleProps` (in the `if (!el.styleProps) {` block), before a default value for `counterProps` is set. This is fixed by assigning a default value for `counterProps` before calling `addProps`.

A test case is also added to reproduce the issue and to prevent future regressions.